Remove Discord bot, dead code, and unused legacy scripts
Four cleanups, all confirmed dead/unused by tracing every caller first:

- Discord support removed entirely: src/discord_bot.py had zero references
  anywhere and can't even be imported (discord isn't a project dependency —
  #180 already had to omit it from coverage for exactly this reason). Also
  removed the "read side" in main.py (_load_discord_state/_apply_discord_change/
  _render_discord_overlay and its call site) since it's useless without a
  working bot, plus the discord_* config keys and README section. The
  config web server (#181) already covers the "change settings remotely"
  use case.

- src/scoreboard_generate.py (1885 lines) deleted. It was a pre-refactor
  monolith duplicating most of fetch_games.py/util.py as its own local
  copies (fetch_scoreboard_for_date, parse_games, check_games_for_sport,
  etc.) — confirmed by diffing against the canonical versions. The only
  thing still used from it was _generate_gif (main.py's auto_generate_video
  timelapse feature), which itself called this file's OWN duplicate
  fetch_scoreboard_for_date rather than fetch_games.py's. Extracted
  _generate_gif and its true dependencies (_fetch_game_timeline,
  _game_state_at_time, _any_game_active, _inning_runs_at_time, _ordinal,
  _parse_mlb_time — none of which have equivalents elsewhere) into a new
  focused src/timelapse.py, rewired to use the canonical fetch_games/util
  functions, with a minimal CLI replacing the old monolithic one (which
  duplicated main.py's entire fetch/render/display/night-mode/Discord
  orchestration just to expose --gif flags).

- Four one-off stadium data-generation scripts removed (extract_mlbam_walls.py,
  fetch_stadium_polygons.py, generate_wall_polygons.py, generate_stadium_svgs.py) —
  undocumented, not referenced by the live pipeline, and the data/SVGs they
  produced are already committed (data/mlbam_walls.json, data/stadium_svgs/).

- Four legacy ad hoc debug scripts in src/ removed (test_emoji.py,
  test_logos.py, test_scoreboard.py, test_stadiums.py) plus
  generate_standings_image.py (confirmed unreferenced dead script) —
  superseded by the real pytest suite.

pyproject.toml's coverage omit list trimmed to just what's still true:
hardware drivers/probes (display.py, display_eink.py, epd_7in5_V2_test.py),
waveshare_epd/*, and download_logos.py (still a real, documented,
infrequently-run tool — kept). timelapse.py gets full test coverage
(99.8%) rather than an omit entry, matching the project's standard for
real pipeline code.

Full suite: 1512 passed, 99.03% coverage (gate: 99%), ruff/mypy clean.
